+++ backend/src/flowers/flowers.module.ts
@@ -1,6 +1,7 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { FlowerLike } from '../likes/flower-like.entity';
+import { FlowerComment } from '../comments/flower-comment.entity';
 import { StorageModule } from '../storage/storage.module';
 import { FlowerPhoto } from './flower-photo.entity';
 import { FlowerPhotosController } from './flower-photos.controller';
@@ -11,7 +12,7 @@ import { FlowersService } from './flowers.service';
 
 @Module({
   imports: [
-    TypeOrmModule.forFeature([Flower, FlowerPhoto, FlowerLike]),
+    TypeOrmModule.forFeature([Flower, FlowerPhoto, FlowerLike, FlowerComment]),
     StorageModule,
   ],
   controllers: [FlowersController, FlowerPhotosController],